feat: merge test into solution
