+++ src/Y2021/D05/Y2021D05.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"bufio"
+	"testing"
+
 	"fmt"
 	"log"
 	"os"
@@ -143,6 +145,19 @@ func BuildDynamicDiagram(size int) [][]int {
 	}
 
 	return diagram
+}
+
+func TestVents(t *testing.T) {
+	tokens := []string{"0,9 -> 5,9", "8,0 -> 0,8", "9,4 -> 3,4", "2,2 -> 2,1", "7,0 -> 7,4",
+		"6,4 -> 2,0", "0,9 -> 2,9", "3,4 -> 1,4", "0,0 -> 8,8", "5,5 -> 8,2"}
+	diagram := BuildDynamicDiagram(10)
+	diagram = AddInputsToDiagram(tokens, diagram)
+	overlaps := NumberOfOverlappingPoints(diagram)
+	PrintDiagram(diagram)
+	fmt.Printf("\nTwo lines overlap on %d points.\n", overlaps)
+	if overlaps != 12 {
+		t.Errorf("Result is incorrect, expected %d, got %d", 12, overlaps)
+	}
 }
 
 func main() {
